refactor(slider): upd export pattern to avoid docgen errors
