+++ packages/components/slider/src/Slider.doc.mdx
@@ -66,13 +66,13 @@ Adheres to the [Slider WAI-ARIA design pattern](https://www.w3.org/WAI/ARIA/apg/
 
 ### Keyboard Interactions
 
-<Kbd>ArrowRight</Kbd>: Increments/decrements by the `step` value.
-<Kbd>ArrowLeft</Kbd>: Increments/decrements by the `step` value.
-<Kbd>ArrowUp</Kbd>: Increases the value by the `step` amount.
-<Kbd>ArrowDown</Kbd>: Decreases the value by the `step` amount.
-<Kbd>PageUp</Kbd>: Increases the value by a larger `step`.
-<Kbd>PageDown</Kbd>: Decreases the value by a larger `step`.
-<Kbd>Shift + ArrowUp</Kbd>: Increases the value by a larger `step`.
-<Kbd>Shift + ArrowDown</Kbd>: Decreases the value by a larger `step`.
-<Kbd>Home</Kbd>: Sets the value to its minimum.
-<Kbd>End</Kbd>: Sets the value to its maximum.
+- <Kbd>ArrowRight</Kbd>: Increments/decrements by the `step` value.
+- <Kbd>ArrowLeft</Kbd>: Increments/decrements by the `step` value.
+- <Kbd>ArrowUp</Kbd>: Increases the value by the `step` amount.
+- <Kbd>ArrowDown</Kbd>: Decreases the value by the `step` amount.
+- <Kbd>PageUp</Kbd>: Increases the value by a larger `step`.
+- <Kbd>PageDown</Kbd>: Decreases the value by a larger `step`.
+- <Kbd>Shift + ArrowUp</Kbd>: Increases the value by a larger `step`.
+- <Kbd>Shift + ArrowDown</Kbd>: Decreases the value by a larger `step`.
+- <Kbd>Home</Kbd>: Sets the value to its minimum.
+- <Kbd>End</Kbd>: Sets the value to its maximum.
