+++ packages/components/slider/src/Slider.styles.ts
