+++ packages/components/slider/src/index.ts
@@ -1,5 +1,15 @@
-export type { SliderRootProps } from './SliderRoot'
-export type { SliderThumb } from './SliderThumb'
-export type { SliderTrack } from './SliderTrack'
+import type { FC } from 'react'
 
-export { Slider } from './Slider'
+import { Slider as Root, type SliderProps } from './Slider'
+import { SliderThumb as Thumb, type SliderThumbProps } from './SliderThumb'
+import { SliderTrack as Track, type SliderTrackProps } from './SliderTrack'
+
+export const Slider: FC<SliderProps> & {
+  Thumb: typeof Thumb
+  Track: typeof Track
+} = Object.assign(Root, {
+  Thumb,
+  Track,
+})
+
+export type { SliderProps, SliderThumbProps, SliderTrackProps }
